[flang] Added tests related to 'num_images()' changes
Summary:
I wrote and added a test to check for semantic errors
in calls to 'num_images()'. I also updated a test,
call10.f90, which has a call to num_images() that previously
produced an error. That line of code no longer
produces an error now that num_images() is evaluted
as intrinsic.